requirements.txt
beautifulsoup4==4.11.1
bing_image_downloader==1.1.2
genanki==0.13.0
googletrans==3.1.0a0
gTTS==2.3.0
Pillow==9.4.0
pydub==0.25.1
pyrubberband==0.3.0
python-dotenv==1.0.0
requests==2.28.1
soundfile==0.11.0
unicode_slugify==0.1.5
validators==0.20.0
wget==3.2
